fastapi_server/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
requests
streamlit

